+++ app/index.tsx
@@ -1,9 +1,14 @@
 import { Link } from "expo-router";
 import { StyleSheet, Text, View } from "react-native";
 
-export default function Page() {
+import PresetList from "../src/components/PresetList";
+
+export default function Homepage(): JSX.Element {
   return (
     <View style={styles.container}>
+      <View style={styles.listContainer}>
+        <PresetList />
+      </View>
       <Link style={styles.button} href="/new-game">
         <Text style={{ color: "#fff" }}>New</Text>
       </Link>
@@ -13,17 +18,19 @@ export default function Page() {
 
 const styles = StyleSheet.create({
   container: {
-    display: "flex",
     backgroundColor: "#333",
-    justifyContent: "flex-end",
-    width: "100%",
-    height: "100%",
-    gap: 8,
     padding: 5,
+    flex: 1,
+    justifyContent: "space-between",
+  },
+  listContainer: {
+    height: "92%",
   },
   button: {
     backgroundColor: "#369",
     padding: 20,
     textAlign: "center",
+    verticalAlign: "middle",
+    height: "8%",
   },
 });