+++ src/types/enums.ts
@@ -1,0 +1,6 @@
+enum CounterModes {
+  INCREMENT,
+  DECREMENT,
+}
+
+export { CounterModes };