+++ src/types/types.ts
@@ -1,0 +1,12 @@
+type Player = {
+  backgroundColor: string;
+};
+
+type Preset = {
+  id: string;
+  name: string;
+  startingLife: number;
+  players: Player[];
+};
+
+export { Player, Preset };